+++ benchmark/insert_benchmark.cc
@@ -35,44 +35,39 @@ enum InsertionType {
 /*
  * Benchmark for adding entries to the index.
  */
-template <dimension_t DIM>
+template <dimension_t DIM, InsertionType TYPE>
 class IndexBenchmark {
+    using Index = PhTree<DIM, int>;
+
   public:
-    IndexBenchmark(
-        benchmark::State& state,
-        TestGenerator data_type,
-        int num_entities,
-        InsertionType insertionType);
+    explicit IndexBenchmark(benchmark::State& state);
 
     void Benchmark(benchmark::State& state);
 
   private:
     void SetupWorld(benchmark::State& state);
 
-    void Insert(benchmark::State& state, PhTree<DIM, int>& tree);
+    void Insert(benchmark::State& state, Index& tree);
 
     const TestGenerator data_type_;
     const size_t num_entities_;
-    const InsertionType insertion_type_;
     std::vector<PhPoint<DIM>> points_;
 };
 
-template <dimension_t DIM>
-IndexBenchmark<DIM>::IndexBenchmark(
-    benchmark::State& state, TestGenerator data_type, int num_entities, InsertionType insertionType)
-: data_type_{data_type}
-, num_entities_(num_entities)
-, insertion_type_(insertionType)
-, points_(num_entities) {
+template <dimension_t DIM, InsertionType TYPE>
+IndexBenchmark<DIM, TYPE>::IndexBenchmark(benchmark::State& state)
+: data_type_{static_cast<TestGenerator>(state.range(1))}
+, num_entities_(state.range(0))
+, points_(state.range(0)) {
     logging::SetupDefaultLogging();
     SetupWorld(state);
 }
 
-template <dimension_t DIM>
-void IndexBenchmark<DIM>::Benchmark(benchmark::State& state) {
+template <dimension_t DIM, InsertionType TYPE>
+void IndexBenchmark<DIM, TYPE>::Benchmark(benchmark::State& state) {
     for (auto _ : state) {
         state.PauseTiming();
-        auto* tree = new PhTree<DIM, int>();
+        auto* tree = new Index();
         state.ResumeTiming();
 
         Insert(state, *tree);
@@ -84,8 +79,8 @@ void IndexBenchmark<DIM>::Benchmark(benchmark::State& state) {
     }
 }
 
-template <dimension_t DIM>
-void IndexBenchmark<DIM>::SetupWorld(benchmark::State& state) {
+template <dimension_t DIM, InsertionType TYPE>
+void IndexBenchmark<DIM, TYPE>::SetupWorld(benchmark::State& state) {
     logging::info("Setting up world with {} entities and {} dimensions.", num_entities_, DIM);
     CreatePointData<DIM>(points_, data_type_, num_entities_, 0, GLOBAL_MAX);
 
@@ -95,9 +90,9 @@ void IndexBenchmark<DIM>::SetupWorld(benchmark::State& state) {
     logging::info("World setup complete.");
 }
 
-template <dimension_t DIM>
-void IndexBenchmark<DIM>::Insert(benchmark::State& state, PhTree<DIM, int>& tree) {
-    switch (insertion_type_) {
+template <dimension_t DIM, InsertionType TYPE>
+void IndexBenchmark<DIM, TYPE>::Insert(benchmark::State& state, Index& tree) {
+    switch (TYPE) {
     case INSERT: {
         for (size_t i = 0; i < num_entities_; ++i) {
             tree.insert(points_[i], (int)i);
@@ -125,71 +120,108 @@ void IndexBenchmark<DIM>::Insert(benchmark::State& state, PhTree<DIM, int>& tree
 }  // namespace
 
 template <typename... Arguments>
-void PhTree3D(benchmark::State& state, Arguments&&... arguments) {
-    IndexBenchmark<3> benchmark{state, arguments...};
+void PhTree3D_INS(benchmark::State& state, Arguments&&...) {
+    IndexBenchmark<3, INSERT> benchmark{state};
+    benchmark.Benchmark(state);
+}
+
+template <typename... Arguments>
+void PhTree3D_EMP(benchmark::State& state, Arguments&&...) {
+    IndexBenchmark<3, EMPLACE> benchmark{state};
+    benchmark.Benchmark(state);
+}
+
+template <typename... Arguments>
+void PhTree3D_SQB(benchmark::State& state, Arguments&&...) {
+    IndexBenchmark<3, SQUARE_BR> benchmark{state};
     benchmark.Benchmark(state);
 }
 
 // index type, scenario name, data_generator, num_entities, function_to_call
-// PhTree 3D CUBE
-BENCHMARK_CAPTURE(PhTree3D, INS_CU_1K, TestGenerator::CUBE, 1000, INSERT)
+BENCHMARK_CAPTURE(PhTree3D_INS, INSERT, 0)
+    ->RangeMultiplier(10)
+    ->Ranges({{1000, 10 * 1000 * 1000}, {TestGenerator::CLUSTER, TestGenerator::CUBE}})
     ->Unit(benchmark::kMillisecond);
 
-BENCHMARK_CAPTURE(PhTree3D, INS_CU_10K, TestGenerator::CUBE, 10000, INSERT)
+BENCHMARK_CAPTURE(PhTree3D_EMP, EMPLACE, 0)
+    ->RangeMultiplier(10)
+    ->Ranges({{1000, 10 * 1000 * 1000}, {TestGenerator::CLUSTER, TestGenerator::CUBE}})
     ->Unit(benchmark::kMillisecond);
 
-BENCHMARK_CAPTURE(PhTree3D, INS_CU_100K, TestGenerator::CUBE, 100000, INSERT)
+BENCHMARK_CAPTURE(PhTree3D_SQB, SQUARE_BR, 0)
+    ->RangeMultiplier(10)
+    ->Ranges({{1000, 10 * 1000 * 1000}, {TestGenerator::CLUSTER, TestGenerator::CUBE}})
     ->Unit(benchmark::kMillisecond);
 
-BENCHMARK_CAPTURE(PhTree3D, INS_CU_1M, TestGenerator::CUBE, 1000000, INSERT)
-    ->Unit(benchmark::kMillisecond);
+// BENCHMARK_CAPTURE(PhTree3D, INSERT, 0)
+//     ->RangeMultiplier(10)
+//     ->Ranges({{1000, 10 * 1000 * 1000}, {TestGenerator::CLUSTER, TestGenerator::CUBE}})
+//     ->Unit(benchmark::kMillisecond);
+//
+// BENCHMARK_CAPTURE(PhTree3D, INSERT, 0)
+//     ->RangeMultiplier(10)
+//     ->Ranges({{1000, 10 * 1000 * 1000}, {TestGenerator::CLUSTER, TestGenerator::CUBE}})
+//     ->Unit(benchmark::kMillisecond);
 
-BENCHMARK_CAPTURE(PhTree3D, INS_CU_10M, TestGenerator::CUBE, 10000000, INSERT)
-    ->Unit(benchmark::kMillisecond);
-
-BENCHMARK_CAPTURE(PhTree3D, EMP_CU_1K, TestGenerator::CUBE, 1000, EMPLACE)
-    ->Unit(benchmark::kMillisecond);
-
-BENCHMARK_CAPTURE(PhTree3D, EMP_CU_10K, TestGenerator::CUBE, 10000, EMPLACE)
-    ->Unit(benchmark::kMillisecond);
-
-BENCHMARK_CAPTURE(PhTree3D, EMP_CU_100K, TestGenerator::CUBE, 100000, EMPLACE)
-    ->Unit(benchmark::kMillisecond);
-
-BENCHMARK_CAPTURE(PhTree3D, EMP_CU_1M, TestGenerator::CUBE, 1000000, EMPLACE)
-    ->Unit(benchmark::kMillisecond);
-
-BENCHMARK_CAPTURE(PhTree3D, EMP_CU_10M, TestGenerator::CUBE, 10000000, EMPLACE)
-    ->Unit(benchmark::kMillisecond);
-
-BENCHMARK_CAPTURE(PhTree3D, SQB_CU_1K, TestGenerator::CUBE, 1000, SQUARE_BR)
-    ->Unit(benchmark::kMillisecond);
-
-BENCHMARK_CAPTURE(PhTree3D, SQB_CU_10K, TestGenerator::CUBE, 10000, SQUARE_BR)
-    ->Unit(benchmark::kMillisecond);
-
-BENCHMARK_CAPTURE(PhTree3D, SQB_CU_100K, TestGenerator::CUBE, 100000, SQUARE_BR)
-    ->Unit(benchmark::kMillisecond);
-
-BENCHMARK_CAPTURE(PhTree3D, SQB_CU_1M, TestGenerator::CUBE, 1000000, SQUARE_BR)
-    ->Unit(benchmark::kMillisecond);
-
-BENCHMARK_CAPTURE(PhTree3D, SQB_CU_10M, TestGenerator::CUBE, 10000000, SQUARE_BR)
-    ->Unit(benchmark::kMillisecond);
-
-BENCHMARK_CAPTURE(PhTree3D, EMP_CL_1K, TestGenerator::CLUSTER, 1000, EMPLACE)
-    ->Unit(benchmark::kMillisecond);
-
-BENCHMARK_CAPTURE(PhTree3D, EMP_CL_10K, TestGenerator::CLUSTER, 10000, EMPLACE)
-    ->Unit(benchmark::kMillisecond);
-
-BENCHMARK_CAPTURE(PhTree3D, EMP_CL_100K, TestGenerator::CLUSTER, 100000, EMPLACE)
-    ->Unit(benchmark::kMillisecond);
-
-BENCHMARK_CAPTURE(PhTree3D, EMP_CL_1M, TestGenerator::CLUSTER, 1000000, EMPLACE)
-    ->Unit(benchmark::kMillisecond);
-
-BENCHMARK_CAPTURE(PhTree3D, EMP_CL_10M, TestGenerator::CLUSTER, 10000000, EMPLACE)
-    ->Unit(benchmark::kMillisecond);
+//// PhTree 3D CUBE
+// BENCHMARK_CAPTURE(PhTree3D, INS_CU_1K, TestGenerator::CUBE, 1000, INSERT)
+//     ->Unit(benchmark::kMillisecond);
+//
+// BENCHMARK_CAPTURE(PhTree3D, INS_CU_10K, TestGenerator::CUBE, 10000, INSERT)
+//     ->Unit(benchmark::kMillisecond);
+//
+// BENCHMARK_CAPTURE(PhTree3D, INS_CU_100K, TestGenerator::CUBE, 100000, INSERT)
+//     ->Unit(benchmark::kMillisecond);
+//
+// BENCHMARK_CAPTURE(PhTree3D, INS_CU_1M, TestGenerator::CUBE, 1000000, INSERT)
+//     ->Unit(benchmark::kMillisecond);
+//
+// BENCHMARK_CAPTURE(PhTree3D, INS_CU_10M, TestGenerator::CUBE, 10000000, INSERT)
+//     ->Unit(benchmark::kMillisecond);
+//
+// BENCHMARK_CAPTURE(PhTree3D, EMP_CU_1K, TestGenerator::CUBE, 1000, EMPLACE)
+//     ->Unit(benchmark::kMillisecond);
+//
+// BENCHMARK_CAPTURE(PhTree3D, EMP_CU_10K, TestGenerator::CUBE, 10000, EMPLACE)
+//     ->Unit(benchmark::kMillisecond);
+//
+// BENCHMARK_CAPTURE(PhTree3D, EMP_CU_100K, TestGenerator::CUBE, 100000, EMPLACE)
+//     ->Unit(benchmark::kMillisecond);
+//
+// BENCHMARK_CAPTURE(PhTree3D, EMP_CU_1M, TestGenerator::CUBE, 1000000, EMPLACE)
+//     ->Unit(benchmark::kMillisecond);
+//
+// BENCHMARK_CAPTURE(PhTree3D, EMP_CU_10M, TestGenerator::CUBE, 10000000, EMPLACE)
+//     ->Unit(benchmark::kMillisecond);
+//
+// BENCHMARK_CAPTURE(PhTree3D, SQB_CU_1K, TestGenerator::CUBE, 1000, SQUARE_BR)
+//     ->Unit(benchmark::kMillisecond);
+//
+// BENCHMARK_CAPTURE(PhTree3D, SQB_CU_10K, TestGenerator::CUBE, 10000, SQUARE_BR)
+//     ->Unit(benchmark::kMillisecond);
+//
+// BENCHMARK_CAPTURE(PhTree3D, SQB_CU_100K, TestGenerator::CUBE, 100000, SQUARE_BR)
+//     ->Unit(benchmark::kMillisecond);
+//
+// BENCHMARK_CAPTURE(PhTree3D, SQB_CU_1M, TestGenerator::CUBE, 1000000, SQUARE_BR)
+//     ->Unit(benchmark::kMillisecond);
+//
+// BENCHMARK_CAPTURE(PhTree3D, SQB_CU_10M, TestGenerator::CUBE, 10000000, SQUARE_BR)
+//     ->Unit(benchmark::kMillisecond);
+//
+// BENCHMARK_CAPTURE(PhTree3D, EMP_CL_1K, TestGenerator::CLUSTER, 1000, EMPLACE)
+//     ->Unit(benchmark::kMillisecond);
+//
+// BENCHMARK_CAPTURE(PhTree3D, EMP_CL_10K, TestGenerator::CLUSTER, 10000, EMPLACE)
+//     ->Unit(benchmark::kMillisecond);
+//
+// BENCHMARK_CAPTURE(PhTree3D, EMP_CL_100K, TestGenerator::CLUSTER, 100000, EMPLACE)
+//     ->Unit(benchmark::kMillisecond);
+//
+// BENCHMARK_CAPTURE(PhTree3D, EMP_CL_1M, TestGenerator::CLUSTER, 1000000, EMPLACE)
+//     ->Unit(benchmark::kMillisecond);
+//
+// BENCHMARK_CAPTURE(PhTree3D, EMP_CL_10M, TestGenerator::CLUSTER, 10000000, EMPLACE)
+//     ->Unit(benchmark::kMillisecond);
 
 BENCHMARK_MAIN();